backend/requirements-hf.txt
# Hugging Face Spaces Optimized Requirements
# Lighter version for faster builds

# FastAPI & Web Framework
fastapi==0.118.0
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.9.1
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Networking
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.1
aiofiles==23.2.1

# Database & ORM
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0

# XML Processing (Tally)
lxml==4.9.3

# Scientific & Data Processing (minimal)
numpy==1.24.3
pandas==2.1.4

# Caching & Utilities
cachetools==5.3.2
python-dateutil==2.8.2
pytz==2023.3

# Logging
python-json-logger==2.0.7

# Optional - Uncomment if needed
# langchain==0.1.0
# langchain-community==0.0.10
# chromadb==0.4.24
# sentence-transformers==2.2.2
# ollama==0.6.1
# PyPDF2==3.0.1
# python-docx==1.1.0
# pillow==10.1.0
# pytesseract==0.3.10

